feat(websockets): make use of websocket hook in HomePage
This adds the additional tab for real time notifications by subscribing
to a websocket that triggers messages on student creation. The user can
subscribe in the Real Time Notification tab. The user can also
unsubscribe but the UI needs some work and correction. Includes a new
useEffect for watching lastMessage from the socket